🚚 continuities move from __unstable__ to core
