+++ packages/atom.io/__tests__/realtime-ipc/BrowserGame.tsx
@@ -1,13 +1,13 @@
 import * as AR from "atom.io/react"
-import { myIdState } from "atom.io/realtime-client"
+import * as RTC from "atom.io/realtime-client"
 import * as RTR from "atom.io/realtime-react"
 import * as RTS from "atom.io/realtime-server"
 import * as React from "react"
-import { useSyncContinuity } from "../../__unstable__/realtime-continuities/use-sync-continuity"
+
 import { gameContinuity, letterAtoms } from "./game-store"
 
 function Room({ roomId }: { roomId: string }): JSX.Element {
-	useSyncContinuity(gameContinuity)
+	RTR.useSyncContinuity(gameContinuity)
 	const letter0 = AR.useO(letterAtoms, 0)
 	return (
 		<main data-testid={roomId}>
@@ -75,6 +75,6 @@ export function A(props: { mySocketKey: string }): JSX.Element | null {
 }
 
 export function BrowserGame(): JSX.Element | null {
-	const mySocketKey = AR.useO(myIdState)
+	const mySocketKey = AR.useO(RTC.myIdState)
 	return mySocketKey ? <A mySocketKey={mySocketKey} /> : null
 }
