+++ packages/atom.io/__tests__/realtime/realtime-sync-transaction.test.tsx
@@ -7,7 +7,7 @@ import * as RTR from "atom.io/realtime-react"
 import * as RTS from "atom.io/realtime-server"
 import * as RTTest from "atom.io/realtime-testing"
 import * as React from "react"
-import { continuity } from "../../__unstable__/realtime-continuities/realtime-continuity"
+import { continuity } from "../../realtime/src/realtime-continuity"
 import { throwUntil } from "../__util__/waiting"
 
 AtomIO.getState(RTC.myIdState)
