+++ packages/atom.io/realtime-client/src/index.ts
@@ -7,5 +7,6 @@ export * from "./pull-selector-family-member"
 export * from "./push-state"
 export * from "./realtime-client-stores"
 export * from "./server-action"
+export * from "./sync-continuity"
 export * from "./sync-server-action"
 export * from "./sync-state"
