+++ packages/atom.io/realtime-react/src/index.ts
@@ -7,5 +7,6 @@ export * from "./use-pull-selector"
 export * from "./use-pull-selector-family-member"
 export * from "./use-push"
 export * from "./use-server-action"
+export * from "./use-sync-continuity"
 export * from "./use-sync-server-action"
 export * from "./use-sync"
