+++ packages/atom.io/realtime-react/src/use-sync-continuity.ts
@@ -1,8 +1,8 @@
 import { StoreContext } from "atom.io/react"
 import * as React from "react"
-import { useRealtimeService } from "../../realtime-react/src/use-realtime-service"
-import type { ContinuityToken } from "./realtime-continuity"
-import { syncContinuity } from "./sync-continuity"
+import { syncContinuity } from "../../realtime-client/src/sync-continuity"
+import type { ContinuityToken } from "../../realtime/src/realtime-continuity"
+import { useRealtimeService } from "./use-realtime-service"
 
 export function useSyncContinuity(token: ContinuityToken): void {
 	const store = React.useContext(StoreContext)
