+++ packages/atom.io/realtime/tsup.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from "tsup"
+import { DTS_OPTIONS } from "~/packages/atom.io/tsup.config"
+
+export default defineConfig(DTS_OPTIONS)
