+++ packages/atom.io/tsconfig.prod.json
@@ -30,6 +30,9 @@
 		"react-devtools/dist",
 		"react-devtools/package.json",
 		"react-devtools/src",
+		"realtime/dist",
+		"realtime/package.json",
+		"realtime/src",
 		"realtime-client/dist",
 		"realtime-client/package.json",
 		"realtime-client/src",
