+++ packages/atom.io/tsup.config.ts
@@ -40,8 +40,9 @@ export const JS_OPTIONS: Options = {
 		"internal/dist/index": `internal/src/index.ts`,
 		"introspection/dist/index": `introspection/src/index.ts`,
 		"json/dist/index": `json/src/index.ts`,
-		"react/dist/index": `react/src/index.ts`,
 		"react-devtools/dist/index": `react-devtools/src/index.ts`,
+		"react/dist/index": `react/src/index.ts`,
+		"realtime/dist/index": `realtime/src/index.ts`,
 		"realtime-client/dist/index": `realtime-client/src/index.ts`,
 		"realtime-react/dist/index": `realtime-react/src/index.ts`,
 		"realtime-server/dist/index": `realtime-server/src/index.ts`,
